WL-4030 Refactored to top-level class.
This just makes the code more readable, rather than having static imports and there was no reason for the class to be an inner class.
