+++ api/src/main/java/uk/ac/ox/oucs/vle/CourseComponentExport.java
@@ -27,26 +27,4 @@ public class CourseComponentExport {
         signups.add(signup);
     }
 
-    /**
-     * Holds the signup and group that were made against a component.
-     * This class exists so that we know we have loaded both objects and don't
-     * have to go back to hibernate when wanting details of the group associated with the signup.
-     */
-    public static class CourseSignupExport {
-        private CourseSignup signup;
-        private CourseGroup group;
-
-        public CourseSignupExport(CourseSignup signup, CourseGroup group) {
-            this.signup = signup;
-            this.group = group;
-        }
-
-        public CourseSignup getSignup() {
-            return signup;
-        }
-
-        public CourseGroup getGroup() {
-            return group;
-        }
-    }
 }
